apps/api/pyproject.toml
[project]
name = "mangomelt-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "sqlmodel==0.0.22",
  "psycopg[binary]==3.2.1",
  "alembic==1.13.2",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "redis==5.0.8",
  "boto3==1.35.36",
  "stripe==10.10.0",
  "httpx==0.27.2",
  "loguru==0.7.2"
]
